redesign of some airfield attributes
git-svn-id: https://svn.kflog.org/svn/repos/cumulus/qt4/trunk@4548 c8f415cb-fa42-0410-b21e-89df95f64ae5

+++ cumulus/airfieldlistwidget.cpp
@@ -108,7 +108,7 @@ Waypoint* AirfieldListWidget::getCurrentWaypoint()
   wp.type = site->getTypeID();
   wp.elevation = site->getElevation();
   wp.icao = site->getICAO();
-  wp.frequency = site->getFrequency().toDouble();
+  wp.frequency = site->getFrequency();
   wp.runway = site->getRunway().direction;
   wp.length = site->getRunway().length;
   wp.surface = site->getRunway().surface;
